v1.6.26: Some bugfixes: errors when emptying db and installing Demo db, added auto redirect to feed

+++ backup_restore.php
@@ -199,17 +199,19 @@ elseif (isset($_REQUEST['empty'])) {
 	get_tags($refresh = 1);
 	get_texttags($refresh = 1);
 	$dir = './thumbnails/' . $tbpref . 'thumbs';
-	$it = new RecursiveDirectoryIterator($dir, RecursiveDirectoryIterator::SKIP_DOTS);
-	$files = new RecursiveIteratorIterator($it,
-	             RecursiveIteratorIterator::CHILD_FIRST);
-	foreach($files as $file) {
-	    if ($file->isDir()){
-	        rmdir($file->getRealPath());
-	    } else {
-	        unlink($file->getRealPath());
-	    }
+	if(is_dir($dir)){
+		$it = new RecursiveDirectoryIterator($dir, RecursiveDirectoryIterator::SKIP_DOTS);
+		$files = new RecursiveIteratorIterator($it,
+			     RecursiveIteratorIterator::CHILD_FIRST);
+		foreach($files as $file) {
+		    if ($file->isDir()){
+			rmdir($file->getRealPath());
+		    } else {
+			unlink($file->getRealPath());
+		    }
+		}
+		rmdir($dir);
 	}
-	rmdir($dir);
 	$message = "Database content has been deleted (but settings have been kept)";
 }
